services/backend/requirements.txt
fastapi==0.129.0
uvicorn[standard]==0.40.0
pymssql==2.3.13
python-dotenv==1.2.1
